requirements.txt
python-dotenv==1.1.1
langchain==0.3.27
langchain-openai==0.3.33
azure-search-documents==11.5.3
azure-core==1.35.1
azure-identity==1.25.0
tree_sitter==0.20.4
atlassian-python-api=4.0.7